backend/requirements.txt
supabase
python-dotenv
pytest
gotrue
fastapi
uvicorn[standard]
pydantic-settings
